feat: Common struct for best market price and outer index count
TODOs
- Use MarketStateV2
- Enforce max tick size in external functions
- Use BestPriceAndIndexCount struct

+++ crates/goblin-market/src/program/processor/new_order/new_order.rs
@@ -108,7 +108,7 @@ pub fn place_multiple_new_orders(
                 }
                 // Price can't exceed max
                 require!(
-                    condensed_order.price_in_ticks <= Ticks::new(MAX_TICK),
+                    condensed_order.price_in_ticks <= Ticks::new(MAX_TICK as u64),
                     GoblinError::PricesNotInOrder(PricesNotInOrder {})
                 );
             }
